Workaround --strict as per https://github.com/krzyzanowskim/OpenSSL/commit/132e40713b3a1f74efbc3b74a547088c7cf703ca#commitcomment-44154022

+++ scripts/build.sh
@@ -243,7 +243,12 @@ if [ ! -f "${SCRIPT_DIR}/../openssl-${OPENSSL_VERSION}.tar.gz" ]; then
    curl -fL "https://www.openssl.org/source/openssl-${OPENSSL_VERSION}.tar.gz" -o "${SCRIPT_DIR}/../openssl-${OPENSSL_VERSION}.tar.gz"
    curl -fL "https://www.openssl.org/source/openssl-${OPENSSL_VERSION}.tar.gz.sha256" -o "${SCRIPT_DIR}/../openssl-${OPENSSL_VERSION}.tar.gz.sha256"
    DIGEST=$( cat ${SCRIPT_DIR}/../openssl-${OPENSSL_VERSION}.tar.gz.sha256 )
-   echo "${DIGEST}  openssl-${OPENSSL_VERSION}.tar.gz" | shasum -a 256 --strict -c -
+
+   if [[ "$(shasum -a 256 "openssl-${OPENSSL_VERSION}.tar.gz" | awk '{print $1}')" != "${DIGEST}" ]]
+   then
+      echo "openssl-${OPENSSL_VERSION}.tar.gz: checksum mismatch"
+      exit 1
+   fi
    rm -f "${SCRIPT_DIR}/../openssl-${OPENSSL_VERSION}.tar.gz.sha256"
 fi
